fix(shard): guard cross-shard deferred response against fd reuse
Tightens the cross-shard blocking-read delivery path with four
related fixes uncovered while reviewing the original change:

- Waiters now carry conn_id; deliverDeferred verifies it against the
  live connection before writing. The owner shard's closeConnection
  only removes waiters from its own pool, so a waiter on a different
  data shard can outlive the connection. Without the check, an
  inbox-borne response could land on a new connection that the kernel
  has assigned the same fd to. conn_id is packed into the inbox
  sequence field — no inbox layout change.

- RegisterOpts.owner_shard / conn_id are required (no defaults). A
  forgotten field silently routed the response to the wrong shard;
  making them non-default turns that into a compile error.

- owner_shard widened from u8 to u16 to match shard_count. The u8
  would have panicked on @intCast at >255 shards.

- Pool-full now produces an immediate fallback response (kv: not_found,
  queue: empty, stream/action: empty OK) at every register() site, not
  just stream_group_read. Previously four sites discarded the return
  value and set response_deferred, leaving the request to hang until
  the client's deadline.

+++ src/node/connection.zig
@@ -313,13 +313,13 @@ pub const Connection = struct {
     /// fd, buffers, and reactor registration). Blocking-read waiters carry
     /// this so a deferred response produced on another (data) shard can be
     /// marshalled back to the owning shard's thread for the socket write.
-    owner_shard: u8,
+    owner_shard: u16,
 
     /// Set by handlers that intentionally defer the response (e.g. blocking GET).
     /// processRequests checks this to suppress the default "not implemented" error.
     response_deferred: bool,
 
-    pub fn init(allocator: std.mem.Allocator, fd: i32, conn_id: u32, owner_shard: u8) !Connection {
+    pub fn init(allocator: std.mem.Allocator, fd: i32, conn_id: u32, owner_shard: u16) !Connection {
         var read_buf = try RingBuffer.init(allocator);
         errdefer read_buf.deinit();
         var write_buf = try RingBuffer.init(allocator);
